+++ langs/extort/Makefile
@@ -7,8 +7,11 @@ else
   format=elf64
 endif
 
-%.run: %.o main.o char.o byte.o
-	gcc main.o char.o byte.o $< -o $@
+%.run: %.o runtime.o
+	gcc runtime.o $< -o $@
+
+runtime.o: main.o char.o io.o io_set.o
+	gcc -r main.o char.o io.o io_set.o -o runtime.o
 
 main.o: main.c types.h
 	gcc -c main.c -o main.o
@@ -16,11 +19,11 @@ main.o: main.c types.h
 char.o: char.c types.h
 	gcc -c char.c -o char.o
 
-byte.o: byte.c types.h
-	gcc -c byte.c -o byte.o
+io.o: io.c io.h
+	gcc -c io.c -o io.o
 
-byte-shared.o: byte-shared.c types.h
-	gcc -fPIC -c byte-shared.c -o byte-shared.o
+io_set.o: io_set.c io.h
+	gcc -c io_set.c -o io_set.o
 
 %.o: %.s
 	nasm -f $(format) -o $@ $<
