+++ langs/extort/compile.rkt
@@ -7,14 +7,19 @@
   (prog (Extern 'peek_byte)
         (Extern 'read_byte)
         (Extern 'write_byte)
+        (Extern 'error)
         (Label 'entry)
         (Sub 'rsp 8)
         (compile-e e)
         (Add 'rsp 8)
         (Ret)
+        ;; Error handler
         (Label 'err)
+        (Lea 'rax 'error)
+        (Mov 'rax (Offset 'rax 0))
+        (Call 'rax)
         (Add 'rsp 8)
-        (Call 'error)))
+        (Ret)))
 
 ;; Expr -> Asm
 (define (compile-e e)
